マイグレーションファイルを修正

+++ db/migrate/20170710071727_remove_dont_use_table.rb
@@ -1,8 +1,8 @@
 class RemoveDontUseTable < ActiveRecord::Migration
   def change
     drop_table :training_texts
-    drop_table :tags
-    drop_table :taggings
+    ActiveRecord::Migration.drop_table :tags if ActiveRecord::Base.connection.table_exists? :tags
+    ActiveRecord::Migration.drop_table :taggings if ActiveRecord::Base.connection.table_exists? :taggings
     drop_table :contexts
     drop_table :contact_states
     drop_table :contact_answers
